Feat: 새로고침 감지
